% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{cmh.pvalues.support}
\alias{cmh.pvalues.support}
\title{Computing p-values and their support for Cochran-Mantel-Haenszel (CMH) 
exact tests}
\usage{
cmh.pvalues.support(counts, alternative = "two.sided")
}
\arguments{
\item{counts}{a matrix of 4xK columns and any number of lines,
each line representing a 2 x 2 x K contingency table to
test. Make sure all 2x2 contingency tables are stored 
in the same order into a 4 element vector, which are 
concatenated to form a row with length of 4xK.}

\item{alternative}{same argument as in \code{\link{fisher.test}}. The three
possible values are \code{"greater"} (default),
\code{"two.sided"} or \code{"less"} and you can specify
just the initial letter.}
}
\value{
A list of two elements:
\item{raw}{raw discrete p-values.}
\item{support}{a list of the supports of the CDFs of the p-values.
Each support is represented by a vector in increasing order.}
}
\description{
Computes discrete raw p-values and their support
for Cochran-Mantel-Haenszel (CMH) exact tests when applied to 
2 x 2 x K contingency tables.
}
\details{
Assume that we have K stratified 2x2 contingency tables resulting from two
random variables, both have only two values. The CMH assumes a common 
odds ratio for all K 2x2 contingency tables. Under the null hypothesis, 
the common odds ratio is 1, i.e., no association between the two random 
variables after stratified by another variable which has K values. 

To calculate the probability of support, the internal R C function C_d2x2xk
used by mantelhaen.test is also used here.

The code for the computation of the p-values of CMH
exact test is inspired by the source code of mantelhaen.test and the function
\code{\link[DiscreteFDR]{fisher.pvalues.support}}.

See the Wikipedia article about Fisher's
exact test, paragraph Example, for a good depiction
of what the code does for each possible value
of \code{alternative}.
}
