% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoCoRV.R
\name{CoCoRV}
\alias{CoCoRV}
\title{Consistent summary Counts based Rare Variant burden test (CoCoRV)}
\usage{
CoCoRV(
  controlGDSFile,
  caseGDSFile,
  controlAnnoGDSFile = "",
  caseAnnoGDSFile = "",
  nControl = -1,
  nCase = -1,
  sampleList = "",
  outputPrefix = "out",
  AFMax = 0.001,
  AFUse = "case_control_joint",
  AFGroup = "max",
  countUse = "sample",
  variantMissing = 0.1,
  removeStar = F,
  minREVEL = 0.65,
  maxAFPopmax = 1,
  ACANConfig = "",
  caseGroup = "",
  variantExcludeFile = "",
  variantIncludeFile = "",
  ignoreFilter = F,
  variantGroupCustom = "",
  extraParamJason = "",
  variantGroup = "pathogenic",
  groupColumn = "Gene.refGene",
  annotationUsed = "",
  bed = "",
  fullCaseGenotype = T,
  overlapType = "within",
  checkHighLDInControl = F,
  gnomADVersion = "v2exome",
  pLDControl = 0.05,
  ORThresholdControl = 1,
  highLDVariantFile = "",
  ignoreEthnicityInLD = T,
  regionFile = "",
  reference = "GRCh37"
)
}
\arguments{
\item{controlGDSFile}{gds format file of controls}

\item{caseGDSFile}{gds format file of cases}

\item{controlAnnoGDSFile}{gds format file of annotated variants in controls}

\item{caseAnnoGDSFile}{gds format file of annotated variants in cases}

\item{nControl}{number of controls, will be detected based on AN if not set}

\item{nCase}{number of cases, will be detected based on AN if not set}

\item{sampleList}{a one column file listing sample IDs used in cases}

\item{outputPrefix}{the output file prefix}

\item{AFMax}{the maximum of the alternate alleles frequency}

\item{AFUse}{how to apply the AF threshold, options are: 
case_control_joint (default), control, case_control_separate}

\item{AFGroup}{how to combine AFs from multiple groups, options are: 
max: use the max of multiple group AFs (default), pool: pool all groups 
together to calculate the AF}

\item{countUse}{how to use the count for association test. 
sample: each sample is a unit in the final association test (default), 
or allele_null: functional allele counts vs nonfunctional allele counts as 
in proxecat}

\item{variantMissing}{the maximum missingness allowed for a variant}

\item{removeStar}{remove overlapping deletion variants with the star symbol.
This is useful when using gnomAD because gnomAD does not have variants with
star as alternate allele.}

\item{minREVEL}{the minimum REVEL score for pathogenic missense variants}

\item{maxAFPopmax}{the maximum AF_popmax annotated from gnomAD, make sure to 
use the same annotated AF_popmax for both the case and the control. 
To do this, first remove the annotation AF_popmax from the raw gnomAD vcf 
using bcftools annotate and then add the annotation from the 
annotation tool, e.g., ANNOVAR}

\item{ACANConfig}{a file specifying the AC AN info ID and optional 
stratified ID for cases}

\item{caseGroup}{a file with sample ID in the first column and the
stratified group info in the 2nd column}

\item{variantExcludeFile}{a one column file specifying the variants to be 
excluded}

\item{variantIncludeFile}{a one column file specifying the variants to be 
included, it will also include variants specified in variantExcludeFile}

\item{ignoreFilter}{if not specified, remove variants with filter not equal 
to PASS in either case or control for a consistent filtering, default is 
false.}

\item{variantGroupCustom}{an optional R file with functions defining the 
variants of interest or a tab separated two column text file where the 
first column is the variantGroup ID and the second column is the filtering
expression using annotation column names.}

\item{extraParamJason}{an optional JASON file to provide extra parameters
to the custuomized file variantGroupCustom}

\item{variantGroup}{a specified variant group to use for test or a self 
defined function to define the variants of interest. See below for more 
details on defined variant sets.}

\item{groupColumn}{the column specifying the column name grouping variants, 
such as the annotated gene name. If NA, all variants will be considered.}

\item{annotationUsed}{a one column file without the header listing 
all required annotations used in variant filtering. AC AN annotations from 
ACANConfig will be added automatically.}

\item{bed}{a bed file to specify the regions considered, e.g., the 
intersection of the bed files from cases and controls, where >= 90% of 
samples have coverage >= 10}

\item{fullCaseGenotype}{If true (default), use the full genotypes to extract 
counts for cases, otherwise summary counts are used}

\item{overlapType}{overlap type between the variant and the bed file: 
"any", "within" (default), "gds"}

\item{checkHighLDInControl}{for each case double het, if high LD in control 
then exclude}

\item{gnomADVersion}{specify the major gnomAD version, v2exome (default),
v2genome, v3genome,
 if gnomAD is used}

\item{pLDControl}{the pvalue threshold to detect high LD variants in control}

\item{ORThresholdControl}{the OR threshold to define high LD variants}

\item{highLDVariantFile}{the precomputed variant pairs in LD in controls}

\item{ignoreEthnicityInLD}{whether to ignore ethnicity when loading the 
precomputed LD pairs}

\item{regionFile}{a bed file with names at the 4th column defining the 
processing unit. This can be used 
to define test regions or to process relatively small regions 
sequentially to reduce memory}

\item{reference}{the reference build, either GRCh37 or GRCh38, used for 
defining the pseudoautosomal region in X and Y}
}
\value{
A list of three components
          testResult: the association test result, 
          caseGeneList: the gene and variants used in the test in cases
          controlGeneList: the gene and variants used in the test in controls
}
\description{
This is the main function for performing variant filtering, count extraction
or estimation, and the rare varaint burden test
}
\details{
Defined variant sets based on annovar annotations
\itemize{
  \item annovar_pathogenic: frameshift_deletion | frameshift_insertion | 
    stopgain | nonsynonymous_SNV with REVEL >= minREVEL   
  \item annovar_function: frameshift_deletion | frameshift_insertion | 
    stopgain | splicing | nonsynonymous_SNV with REVEL >= minREVEL
  \item annovar_missense: nonsynonymous_SNV with REVEL >= minREVEL
  \item annovar_LOF: frameshift_deletion | frameshift_insertion | stopgain
  \item annovar_synonym: synonymous_SNV
  \item annovar_splicing: splicing
  \item annovar_intron: intronic | ncRNA_intronic
  \item annovar_UTR3: UTR3
  \item annovar_UTR5: UTR5
  \item annovar_UTR: UTR3 | UTR5  
}
For a self defined function, it must take a data frame of summary count data
as the parameter and return a logical index on rows, see customVariantSets
in variantSets.R for an example. 

If stratified analysis is specified, the AC and AN will still be used for
filtering and QC, then the counts corresponding to the stratified groups
will be used in the CMH exact test
}
