% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFunctions.R
\name{contingencyTableTest}
\alias{contingencyTableTest}
\title{Perform contingency table based tests}
\usage{
contingencyTableTest(countMatrix, test = "FET")
}
\arguments{
\item{countMatrix}{a matrix of 4 * nGroup, ordered as
caseWithMutation caseWOMutation controlWithMutation controlWOMutation
and then by groups}

\item{test}{name of the test. "FET" means Fisher exact test, "CMH" means CMH}
}
\value{
a matrix of four columns: p-value, odds ratio, 
  95% confidence interval lower bound,
  95% confidence interval upper bound
}
\description{
If the number of columns is 4, Fisher exact test is used
If the number of columns is a multiple of 4 (> 4), CMH exact test is used
}
