% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR.R
\name{countBasedFDR}
\alias{countBasedFDR}
\title{calculate RBH based FDR for FET using a 2x2 table or CMH using 2x2xm table
If the column number of the input is 4, other count based FDRs for FET is 
also calculated}
\usage{
countBasedFDR(
  data,
  alternative = "two.sided",
  nReplication = 1000,
  ncore = 16,
  sampling = "cdf"
)
}
\arguments{
\item{data}{a matrix of 4*n}

\item{alternative}{should be two.sided, greater, or less as in fisher.test}

\item{nReplication}{the number of replications in the resampling based 
method}

\item{ncore}{the number of cores to use}

\item{sampling}{cdf: sampling p-values directly from the cdf of the p-values; 
table: sampling contingency tables first and then calcualte
p-values.}
}
\value{
a matrix of adjusted p-values. If the input is for CMH, then two 
methods are returned: RBH_point_estimate, RBH_upper_limit. If the 
input is for FET, additoinal FET based FDR are also returned.
}
\description{
calculate RBH based FDR for FET using a 2x2 table or CMH using 2x2xm table
If the column number of the input is 4, other count based FDRs for FET is 
also calculated
}
