% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFunctions.R
\name{estimateCounts}
\alias{estimateCounts}
\title{estimate the counts of samples of different models based on AFs of variants}
\usage{
estimateCounts(
  nSample,
  AF,
  excludeHighLD = T,
  highLDList = list(NULL),
  countHomAlt = NULL,
  twoHaps = F
)
}
\arguments{
\item{nSample}{The total number of samples}

\item{AF}{The vector of alternate allele frequencies of each variant}

\item{excludeHighLD}{If true, exclude variants in high LD except the first
variant with the largest AF when calculating the counts, otherwise 
exclude all variants in high LD when calculating the counts and then 
later add the counts due to the first variant of each LD group. 
The default is T, so the variants with high LD are not considered in the 
recessive or 2hets model.}

\item{highLDList}{A list of variants with high LD. Default NULL}

\item{countHomAlt}{Either NULL or a vector of the counts of the alternate
homozygous genotypes. If not null, the estimation is based on 
genotypes instead of assuming Hardy Weinberg equilibrium}

\item{twoHaps}{If true, two variants must be on the two different haplotypes
to be considered as double heterozygous. Default F}
}
\value{
A vector of three integer numbers correspond to the number of 
qualified samples under the dominant model, the recessive and the double
heterozygous model
}
\description{
This function estimates the count of samples under the dominant model, the
recessive model and the double heterozygous model
}
