% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFunctions.R
\name{extractACANAnnotations}
\alias{extractACANAnnotations}
\title{This function extracts annotations from a gds File}
\usage{
extractACANAnnotations(
  gds,
  sampleID = NULL,
  bedGRange = NULL,
  updateACAN = F,
  overlapType = "within",
  ACANIDs = NULL,
  annotationSet = NULL
)
}
\arguments{
\item{gds}{a gds file handle}

\item{sampleID}{If set, then will restrict to the samples specified first}

\item{bedGRange}{If set, it is a GRanges or GRangesList to specify a region}

\item{updateACAN}{whether to calculate the AC AN instead of extracting from
the annotations}

\item{overlapType}{the overlap type between the bedGRange and the variants
in the gdsFile. See type in the findOverlaps function in GenomicRanges}
}
\value{
a matrix with variantID as the first column and the rest are 
annotations in the gds file
}
\description{
This function extracts annotations from a gds File
}
