% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFunctions.R
\name{extractIndependentGnomADCounts}
\alias{extractIndependentGnomADCounts}
\title{extract the independent counts from gnomAD counts}
\usage{
extractIndependentGnomADCounts(
  controlDataPerGene,
  ethnicitySet,
  gnomADVersion = "v2exome"
)
}
\arguments{
\item{controlDataPerGene}{a matrix, where the first column is the variant ID, 
and the rest are gnomAD annotated counts.}

\item{ethnicitySet}{which ethnicities to use in the extraction. For example, 
the 6 major ethnicities: c("nfe", "afr", "amr", "eas", "sas", "fin")}

\item{gnomADVersion}{gnomAD major version (v2exome, v2genome or v3genome), 
default is v2exome.}
}
\value{
a list of two components. The first component is a matrix, where
each column is a vairant, and rows are independent set of summary counts. 
The second component is a vector of averaged ANs across the variants, 
representing an estimate of the total number of haplotyppes pooled.
}
\description{
extract the independent counts from gnomAD counts
}
