% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{Perform Gene set enrichment analysis using simulated null P values 
corresponding to phenotype permutation}
\usage{
GSEA(
  data,
  pathway,
  alternative = "two.sided",
  p = 1,
  minPathwaySize = 2,
  maxPathwaySize = Inf,
  nReplication = 1000,
  maxPermutation = 1e+06,
  maxExtreme = 36,
  ncore = 1
)
}
\arguments{
\item{data}{a  matrix or data frame with the number of columns being 4*k:
If k is 1, it uses the Fisher exact test;
if k > 1, it uses the CMH exact test. The columns should be arranged as
follows for each stratified group
  counts with rare mutations in case, 
  counts without rare mutations in case, 
  counts with rare mutations in controls
  counts without rare mutations in controls}

\item{pathway}{a list of vectors with gene IDs for each pathway}

\item{alternative}{"two.sided", "greater", or "less". The later GSEA will 
match the direction to perform standard GSEA, or one-sided GSEA.}

\item{p}{specify the power of the absolute value of the score, default 1.}

\item{nReplication}{number of actual replications of the null P values, 
these will be used in later sampling to form permuted values in GSEA}

\item{maxPermutation}{number of maximal permutations performed if the 
maximal extreme value is not reached.}

\item{maxExtreme}{number of the maximal extreme values from the permutation 
before stop, used to control the adaptive permutation}
}
\value{
A data frame of GSEA results
}
\description{
Perform Gene set enrichment analysis using simulated null P values 
corresponding to phenotype permutation
}
