% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHighLD.R
\name{LDTest}
\alias{LDTest}
\title{This function uses a set of summary counts to estimate/test the linkage 
disequilibrium}
\usage{
LDTest(
  data,
  ORThreshold = 1,
  alternative = "greater",
  genotype = F,
  ACThreshold = 1
)
}
\arguments{
\item{data}{a matrix of 3 or 2 columns. The first two columns are the pooled
alternate allele counts of the two variants. The third column is the number
of haplotypes in total. If genotype data is used, the third column can be 
omitted}

\item{ORThreshold}{the threshold used in the test}

\item{alternative}{either "greater" or "less". If "greater", the alternative
hypothesis is that the odds ratio theta > ORThreshold. If "less", then it is
theta < ORThreshold}

\item{genotype}{If true, then assume the data is generated using full 
genotype data, i.e., n = 2 for all samples. Note that in this case, each row
corresponds to an individual}

\item{ACThreshold}{the minimal alternate allele count for odds ratio 
estimation or test}
}
\value{
A vector of 12 values: pvalue of the 
likelihood ratio test and the estimated odds ratio, r2, D', log likelihood 
of the full and the null model, the next four are estimated 
probalities/frequencies of the 4 haplotypes: p11, p10, p01, p00, and the 
return status of the optimization for the full and the null model, where
1-4 indicate success, other values indicate failing to converge
}
\description{
It can be applied to both the summary counts or full genotypes, where the
latter is a special case of pooling two haplotypes together
}
