% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHighLDVariants.R
\name{loadHighLDVariants}
\alias{loadHighLDVariants}
\title{load the variants in high LD. 
If the input file is not a rds file, the input file must have the 
following columns:
variant1  variant2  gene  ethnicity pvalue log10OR fdrEthnicity. If the 
input is a rds file, it will be loaded directly and other arguments will
not be applied}
\usage{
loadHighLDVariants(
  highLDVariantFile,
  ignoreEthnicityInLD = F,
  pThreshold = 0.05,
  ORThreshold = 1
)
}
\arguments{
\item{highLDVariantFile:}{a file containing the variants in high LD}

\item{ignoreEthnicityInLD:}{If true, ignore the ethnicity and merge all LDs in any}

\item{ethnicity.}{Otherwise LDs are stored for each ethnicity}

\item{pThreshold:}{the p-value threshold for LD pairs}

\item{ORThreshold:}{the odds ratio threshold for LD pairs}
}
\value{
It returns a list of two compoents. The first one is based on 
the FDR threshold 0.05, and the second one is based on a more lenient 
 pvalue threshold specified by the argument pThreshold. Each component is also a list, where each element corresponds to a gene.
}
\description{
load the variants in high LD. 
If the input file is not a rds file, the input file must have the 
following columns:
variant1  variant2  gene  ethnicity pvalue log10OR fdrEthnicity. If the 
input is a rds file, it will be loaded directly and other arguments will
not be applied
}
\details{
For each gene, if ignoreEthnicityInLD is false, each gene has lists 
corresponding to ethnicities. For example, if ignoreEthnicityInLD is true, the 
highLDList for gene TAAR2 can be accessed as highLDList[["TAAR2"]]. If 
ignoreEthnicityInLD is false, it can be accessed as highLDList[["TAAR2"]][["afr"]] or 
highLDList[["TAAR2"]][["amr"]]
}
