% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplotHGAndFDR}
\alias{qqplotHGAndFDR}
\title{qqplot using the simulated null P values and calculate the FDR designed for
discrete counts}
\usage{
qqplotHGAndFDR(
  data,
  quantile = 0.95,
  nReplication = 1000,
  returnFDR = T,
  returnSimulated = F,
  returnNullP = F,
  maxLogP = 30,
  nullP = NULL,
  doPlot = T,
  limit = NULL,
  seed = NULL,
  plotNULL = F,
  cexRatio = 1.5,
  verbose = T,
  nolambda = F,
  alternative = "two.sided",
  ncore = 1,
  sampling = "cdf",
  ...
)
}
\arguments{
\item{data}{a  matrix or data frame with the number of columns being 4*k:
If k is 1, it uses the Fisher exact test;
if k > 1, it uses the CMH exact test. The columns should be arranged as
follows for each stratified group
  counts with rare mutations in case, 
  counts without rare mutations in case, 
  counts with rare mutations in controls
  counts without rare mutations in controls}

\item{quantile}{the upper quantil of points to use for estimation}

\item{nReplication}{the number of replications of empirical null P values}

\item{returnSimulated}{whether to return simulated null count tables}

\item{returnNullP}{whether to return simulated null P values}

\item{maxLogP}{the maximum of -log10(P) to be plotted, larger values will be 
set to this maximum value}

\item{nullP}{the null P values. If not supplied, it will be simulated}

\item{doPlot}{whether to generate the QQ plot}

\item{limit}{set the x and y axis limit}

\item{seed}{set the random seed}

\item{plotNULL}{whether to plot the boxplot of lambda values under the null}

\item{cexRatio}{set cex values in the plot}

\item{verbose}{whether verbose or not}

\item{nolambda}{if true, no lambda will be estimated/plotted}

\item{alternative}{see alternative in fisher.test}

\item{ncore}{the number of cores used}

\item{sampling}{cdf: sampling p-values directly from the cdf of the p-values; 
table: sampling contingency tables first and then calculate
p-values.}

\item{...}{additional parameters to plot}
}
\value{
A list of the following components. 
lambdaEmp: the estimated inflation factor of the observed data  
lambdaNull: the estimated inflation factors of simulated null data
nullPSortedMean: the mean of sorted null P values
simulatedCount: the simulated count matrix under the null hypothesis
nullP: the matrix of null P-values from the simulated counts
pvalues: the observed P-values
FDRResult: the FDR results if returnFDR is true
indexKept: the index of rows after filtering rows with the total counts 
  being 0 
intercept: the estimated intercept in the QQ plot
}
\description{
If the number of columns is 4, it uses the Fisher exact test.
If the number of columns if a multiple of 4, it uses the CMH exact test
}
