% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplotQT}
\alias{qqplotQT}
\title{QQ plot assuming uniform distribution of the P values}
\usage{
qqplotQT(
  pvalues,
  doPlot = T,
  upper = 0.95,
  lower = 0.5,
  removeOne = F,
  twoPoints = F,
  cexRatio = 1.5,
  maxLogP = 30
)
}
\arguments{
\item{pvalues}{the pvalues used to estimate lambda and generate the QQ plot}

\item{doPlot}{whether to generate the QQ plot}

\item{upper}{the upper quantile of points used}

\item{lower}{the lower quantile of points used}

\item{removeOne}{if true, remove all points with observed P value 1 except 
the point with the highest rank of expected values}

\item{twoPoints}{if ture, the lambda is estimated based on the point with 
the observed pvalue of 1 and the highest rank of expected p-values and 
the pvalue at the 95 percentile among the non-one}

\item{cexRatio}{set the cex option in the plot}

\item{maxLogP}{the maximum of -log10(P) to be plotted, larger values will be
set to this value}
}
\value{
the estimated lambda
}
\description{
It implements 3 different ways to estimate lambda
1) twoPoints: based on the point with the observed pvalue of 1 with the 
    highest rank and the pvalue at the 95 percentile among the non-one 
    pvalues. Set twoPoints to TRUE to run this method
2) 0.5-0.95: fit a line to the points within 0.5 and 0.95 quantiles.
    Set removeOne and twoPoints to FALSE to run this method
3) removeOne: fit a line to the points starting with the point with 
  the observed pvalue of 1 and the highest rank and all points with 
  non-one pvalues until the point with the p-value of 0.95 quantile 
    Set removeOne to TRUE and twoPoints to FALSE to run this method. 
Reference: Gao et al, AJHG, 2018, Burden Testing of Rare Variants Identified 
through Exome Sequencing via Publicly Available Control Data.
}
