% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR.R
\name{RBH}
\alias{RBH}
\title{This function calculates the re-sampling based FDR 
It is from Yekutieli & Benjamini, 1999, but with a minor modification 
on the estimated v(p) under the null, instead of m * p, it is directly 
estimated from the simulated null pvalues. The upper limit RBH only 
considers x = p in Equation 10 of Yekutieli & Benjamini, 1999. 
These modifications are also used in the R package FDR-AME 
implementing these methods, however, they use < instead of <= when defining
r, different from the paper. Here we use <= to define r as in the paper.}
\usage{
RBH(observed, simulated, alpha = 0.05)
}
\arguments{
\item{observed}{the observed p value}

\item{simulated}{a matrix of mTest * nPermutation of p values under the null, 
make sure the rows are consistent between the observed p values and the
simulated p-values}

\item{alpha}{set the 1 - alpha quantile used in the algorithm}
}
\value{
a matrix of two columns, the first is the point estimate FDR, 
the second is the upper limit estimate FDR.
}
\description{
This function calculates the re-sampling based FDR 
It is from Yekutieli & Benjamini, 1999, but with a minor modification 
on the estimated v(p) under the null, instead of m * p, it is directly 
estimated from the simulated null pvalues. The upper limit RBH only 
considers x = p in Equation 10 of Yekutieli & Benjamini, 1999. 
These modifications are also used in the R package FDR-AME 
implementing these methods, however, they use < instead of <= when defining
r, different from the paper. Here we use <= to define r as in the paper.
}
