% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateNullPvalues.R
\name{simulateNullPvalues}
\alias{simulateNullPvalues}
\title{simulated null P values for both the FET and CMH through simulating
the actual 2*2*m tables under the null}
\usage{
simulateNullPvalues(
  data,
  alternative = "two.sided",
  nReplication = 100,
  returnSimulated = F,
  returnNullP = T,
  verbose = T,
  seed = NULL,
  ncore = 1
)
}
\arguments{
\item{data}{a  matrix or data frame with the number of columns being 4*k:
If k is 1, it uses the Fisher exact test;
if k > 1, it uses the CMH exact test. The columns should be arranged as
follows for each stratified group
  counts with rare mutations in case, 
  counts without rare mutations in case, 
  counts with rare mutations in controls
  counts without rare mutations in controls}

\item{alternative}{see alternative in fisher.test}

\item{nReplication}{the number of replications of empirical null P values}

\item{returnSimulated}{whether to return simulated null count tables}

\item{returnNullP}{whether to return simulated null p values}

\item{verbose}{whether verbose or not}

\item{seed}{set the random seed}
}
\value{
A list of the following components. 
simulated: the simulated count matrix under the null hypothesis
nullP: the matrix of null P values from the simulated counts
pvalues: observed p values
}
\description{
If the number of columns is 4, it uses the Fisher exact test.
If the number of columns if a multiple of 4, it uses the CMH exact test
}
